backend/pyproject.toml
[project]
name = "social-knowledge"
version = "0.1.0"
description = "Automated news crawler with AI summarization and Discord notifications"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.10",
    "lxml>=4.9.0",
    "python-dateutil>=2.8.0",
    "google-generativeai>=0.3.0",
    "discord.py>=2.3.0",
    "apscheduler>=3.10.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "firebase-admin>=6.4.0",
    "python-jose>=3.3.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = ["pytest>=7.4.0", "black>=23.0.0", "ruff>=0.1.0"]
